changement sécuritée

+++ commands/conversation/Clearchennel.js
@@ -1,9 +1,10 @@
-const { MessageEmbed } = require('discord.js')
+const { MessageEmbed, Permissions } = require('discord.js')
 
 module.exports = {
 	name: 'clearchannel',
 	description: 'Permet de vider un salon',
 	run: async (client, message, args) => {
+		if (!message.author.permissions.has(Permissions.FLAGS.MANAGE_CHANNELS)) return message.reply({ content: 'Vous n\'avez pas la permission d\'utiliser cette commande.' });
 		await message.channel.delete().then(async channel => {
 			channel.clone(this.name).then(async chann => {
 				chann.send( { embeds: [new MessageEmbed()
@@ -25,6 +26,7 @@ module.exports = {
 	},
 	runSlash: async (client, interaction) => {
 		await interaction.channel.delete().then(async channel => {
+			if (!interaction.user.permissions.has(Permissions.FLAGS.MANAGE_CHANNELS)) return interaction.reply({ content: 'Vous n\'avez pas la permission d\'utiliser cette commande.', ephemeral: true });
 			channel.clone(this.name).then(async chann => {
 				chann.send( { embeds: [new MessageEmbed()
 					.setColor('#4ed5f8')
